requirements.txt
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.0.16
# --- 新增和更新的库 ---
langchain-openai==0.1.0
langchain-huggingface==0.0.1

faiss-cpu==1.7.4
pandas==2.0.0
numpy==1.24.0
requests==2.31.0
python-dotenv==1.0.0
tiktoken==0.5.0
streamlit==1.29.0
